backend/python_scripts/pyproject.toml
[build-system]
requires = ["setuptools>=68.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "argo-energy-backend"
version = "0.2.0"
description = "Argo Energy Solutions — data pipeline and analytics backend"
requires-python = ">=3.9"
dependencies = [
    "psycopg2-binary==2.9.11",
    "SQLAlchemy==2.0.46",
    "requests==2.32.5",
    "pandas==2.3.3",
    "numpy==2.0.2",
    "python-dotenv==1.2.1",
    "python-dateutil==2.9.0.post0",
    "fastapi==0.128.5",
    "uvicorn[standard]==0.39.0",
    "scipy==1.13.1",
    "matplotlib==3.9.4",
    "pytz==2025.2",
    "sentry-sdk==2.52.0",
]

[project.optional-dependencies]
test = [
    "pytest>=8.0",
    "pytest-cov>=5.0",
]

[tool.setuptools.packages.find]
where = ["."]

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]
python_functions = ["test_*"]
